Break out shared CPU code into centralised location
