+++ baseline/integrator/integrator.cc
@@ -22,10 +22,10 @@
 #include "arg_parser.hpp"
 #include "background.cc"
 #include "common.hpp"
-#include "extent.cc"
 #include "ffs_logger.hpp"
 #include "h5read.h"
 #include "integration_calculations.cc"
+#include "integrator/extent.hpp"
 #include "kabsch.cc"
 #include "math/math_utils.cuh"
 #include "predict.cc"
