+++ integrator/CMakeLists.txt
@@ -7,7 +7,7 @@ find_package(Bitshuffle REQUIRED)
 add_executable(integrator
     integrator.cc
     integrator.cu
-    extent.cc
+    ${CMAKE_SOURCE_DIR}/src/integrator/extent.cc
     ../spotfinder/shmread.cc
     ../spotfinder/cbfread.cc
 )
