+++ integrator/integrator.cc
@@ -34,9 +34,9 @@
 #include "common.hpp"
 #include "cuda_arg_parser.hpp"
 #include "cuda_common.hpp"
-#include "extent.hpp"
 #include "ffs_logger.hpp"
 #include "h5read.h"
+#include "integrator/extent.hpp"
 #include "math/math_utils.cuh"
 #include "math/vector3d.cuh"
 #include "predict.cc"
